build: Update the list of headers in srcdir
The generated notify-features.h header lives in builddir, which means it
does not exist in srcdir, where glib-mkenums will look. It also contains
pre-processor symbols only, so it's not really useful.

As we removed it from the list of headers to be scanned by the
introspection parser, we also need to update the files rules for that.