notify-send: allow passing colon as hint value
While some hints like image-path or sound-file supports parameters using the
file:// protocol, it's impossible to pass such hints to notify-send, because
it splits the passed argument so that it can have an unlimited number of colons,
while only 3 should be there (as per TYPE:NAME:VALUE format).

So, make `g_strsplit` to accept only 3 tokens as maximum, while pass the
remaining parameters all-together.

So now it's possible to do
  notify-send -h string:image-path:file:///usr/share/icons/foo.png