notification: Use Desktop Portal notification when running in sandbox
When an application is running from a snap or a flatpak, libnotify
should just work as a FDO Portal Notification wrapper.

As per this, make the library to try use the portal API in sandboxed
applications, by emulating the needed missing pieces.

Not everything can be achieved 1:1, but it's just safer to use in such
contexts.