Set the icon-name property as image hint instead of as app-icon
The icon specified in the FDO notify DBus call isn't an image that maybe
is displayed in a notification but actually the app-icon to use. This
isn't an issue if the notification server displays only one icon/image
but if the app-icon is show in addition to the image in the image hint
this causes obviously a problem, since the server expects the icon to be
a app icon and not an image.