notify-send: Flush stdout after printing notification ID
Consider the following command, priting the ID of a
notification, then waiting for it to close; passing the
output to a script (represented here as `cat`):

    notify-send -w -p foo | cat

Before this patch, running the following command would make it
print no output until the notification was closed.

After this patch, the command properly prints the ID right away,
then exits later when the notification is closed.

This is due to no flush being done after writing the ID, and
stdout not being line-buffered by default when not in TTYs.

This change is useful for any script that wishes to create a
notification, get its ID (for example to manyually dismiss it),
then also know if/when it was closed by the user.