Revert "meson: update stylesheet url"
The change is incorrect.

`http://docbook.sourceforge.net/release/xsl/current/manpages/docbook.xsl` is the variant of the stylesheet that does not use namespaces and will produce a broken manual page. https://gitlab.gnome.org/GNOME/libnotify/-/commit/162d1958694822658dfa27e3dd1c702dee4c285c expects DocBook stylesheet with namespace support.

It does not really matter if the URI ceases resolving to a file – we disable network access for `xsltproc` so it only serves as an identifier for the stylesheet to be discovered locally using a XML catalog.

This reverts commit 2f99025b7ad54f29dc5236aa7dfcfa97d1c8efde.