notify-send: Add an option to set the app-icon
In the default behavior we don't send to the notification any app-icon,
but this may not work well in some implementations, at the same time we
do not want to send as the app-icon what should only be the notification
icon.

So, let's add another app-icon parameter that can be used to specify the
application icon if it's required.

We do not set anything by default though, as it would imply showing the
same icon in proper implementations (as GNOME shell is)