notification: Simplify the logic to support the legacy hints names
As per recent changes we were checking the current spec version in order
to select the hint in use.

However, this was potentially leading to missing icons because at the
moment we were checking for the spec version, the proxy was not
connected yet, and thus such value was always 0.

This was fixed in a previous commit, but we also caused to start a dbus
connection before that was actually needed, or expected (potentially
even during construction).

We prefer some kind of lazy-loading here, and so let's only do the spec
check after that the proxy is already set, at show time.

We can now adjust the hints names during this phase, since it's not
really an expensive operation

Improves: https://gitlab.gnome.org/GNOME/libnotify/-/issues/55