notification: Hold actions in a simpler array
We won't ever have so many action entries to justify an hash-table, and
the duplication removal happened as part of commit 8a7c701f4bd69 broke
the code by changing the order of the items.

So let's just go simpler and keep all the actions in a GPtrArray that is
quick enough to iterate for this purpose.

It also simplifies some cases in which the strictly cased comparson did
not serve us well. In fact now all the actions are case insensitive.